Expproc, Sendproc, and Recvproc are 64-bit (#1016)
Only the DDR versions so far. Made patch for external
testing (Thanks Brian Antalek). Fixed compiler warnings.

+++ src/vnmr/comm.c
@@ -47,8 +47,8 @@ extern int getStatAcqSample();
 extern int getStatLSDV();
 extern int getStatConsoleID();
 extern void expStatusRelease();
-extern long getStatRemainingTime();
-extern long getStatInQue();
+extern int getStatRemainingTime();
+extern int getStatInQue();
 
 #undef DEBUG
